pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "context_engine"
version = "1.0.0"
description = "AI-powered project mapping for the DebugFlow ecosystem"
readme = "README.md"
requires-python = ">=3.8"
dependencies = [
    "tqdm",
    "llama-cpp-python",
     "debugflow", 
]

[project.scripts]
context-engine = "context_engine.cli:main"
context-logs = "context_engine.cli:toggle_logs"
context-logs-on = "context_engine.cli:logs_on"
context-logs-off = "context_engine.cli:logs_off"


[tool.setuptools.packages.find]
where = ["."]
include = ["context_engine*"]